Introduce fileid mapping file
This is an alternate source of the file<->File ID mapping.  It does
not interfere with the use of File IDs embedded in files.

This is meant to allow the use of file-specific dictionaries in a
project without requiring the addition of the scspell-id: tag into its
files.

Given an --override-dictionary=foo, the fileid mapping is stored in
foo.fileid.json, as a mapping from File ID to a list of filenames.  I
chose this arrangment (vs. filename->File ID) to make it easier to see
files that are sharing a fileid.

The filenames are stored as relative paths.  Absolute paths wouldn't
work very well for a dictionary checked into a project, and checked
out in different developer's working directories.  The directory
they're relative to is specified via the --relative-to commandline
argument.  (I expect that a project using this feature will have the
scspell invocation wrapped up in a script that'll provide the
--override-dictionary and --relative-to arguments so that this won't
be onerous.)

This has the disadvantage that the file-specific dictionary won't
"follow" the files accross renames and copies.  scspell.py options
will be introduced in subsequent commits to allow some basic
manipulations without editing the .json file.

I stored the mapping to a separate .json file primarily to take
advantage of json.dump/json.load dealing with spaces and newlines in
filenames.  As it happens, this leaves the regular dictionary file
backwards compatible with old versions of scspell.  scspell won't be
able to find the per-file dictionary for files specified there, but it
will otherwise be able to read and process the dictionary file.

Also change the generation of the handle_add() menu to generate each
optional menu item separately, instead of having a separate menu
hardcoded for each combination.