Store fileid mappings sorted by filename
It's more reader-friendly and easier to edit this way.  And if the
file's being checked into source control, there'll be less churn.  (A
plain dict can be serialized in a completely different order each
time.)

Building the fileid map sorted by filename required a filename->id
dict, which I wanted for regular lookups anyway, so make that part of
the corpus object and maintain it on updates, rather than just
building it in place for the json.dumps().