Support for base dicts
One or more --base-dict's may be specified on the command line.  We
search for tokens in these dicts, but do not add words to them.

--use-builtin-base-dict uses the dictionary.txt in SCSPELL_DATA_DIR as
a base dict.

This allows a project-specific dictionary to be maintained containing
only words from the project, without including the full 844K base
dictionary and its license terms.  E.g.

$ cd PROJECT
$ scspell --override-dictionary .scspell/dictionary.txt \
          --use-builtin-base-dict \
          $PROJECT_FILES

Another benefit is that as scspell's dictionary is updated over time,
scspell invoked with --use-builtin-base-dict will reflect the updated
dictionary automatically.

Implementation notes:

The CorporaFile object for the main dictionary of a run is built with
a list of additional CorporaFile objects, one per base dict.

The list is in the order the base dicts were specified on the command
line.  The order doesn't matter -- the return value of match() depends
only on whether the word is in any of the dictionaries or in none of
them.