Add an osx run to travis
This takes a bit of hacking around, as travis doesn't officially
support python progjects on osx: travis-ci/travis-ci/issues/2312

Copy the approach used in ssokolow/get_user_headers/issues/3

This only runs one version of python, because it ends up with whatever
system python happens to be installed on the osx worker.

The osx workers require "sudo" on the pip install lines.