Exit with status code 2 when user hits ^C
This allows a script that calls scspell to distinguish between scspell
finding a mispelled word, and the user trying to interrupt the script.

e.g.
	scspell file1
	if [ $? == 2 ]; then
	   exit 1
	fi
	scspell file2
	...

Remove the "User abort" prints.  Most programs just exit when the user
hits ^C; there's no reason to do otherwise.

Update the README to document the exit codes.