Added directory traversal
This commit adds the possibility of traversing directories with scspell. Also, the command line argument `--ext` has been added and can be used more than once and all the files in the directories with those extensions will be checked. In case, the argument is not supplied, *all* the files in the directory and *subdirectories* would be checked.

Example:
```scspell dir --ext py --ext txt```

Closes https://github.com/myint/scspell/issues/22